requirements.txt
numpy
tensorflow-cpu
numpy